empty signer is now translated to a 20 byte of 0s fix bug in validate_block_results that didn't pass the block proposer to the tx set. move trigger contract to dir without _ remove all comments from trigger.
